Reminders list added

+++ app/src/main/java/ua/mrrobot1413/movies/domain/useCase/detailed/RemoveFromFavoriteUseCase.kt
@@ -2,18 +2,15 @@ package ua.mrrobot1413.movies.domain.useCase.detailed
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import ua.mrrobot1413.movies.data.network.model.RequestType
-import ua.mrrobot1413.movies.data.storage.model.DetailedMovie
-import ua.mrrobot1413.movies.data.storage.model.FavoriteMovie
-import ua.mrrobot1413.movies.data.storage.model.Genre
-import ua.mrrobot1413.movies.domain.DetailedRepository
-import ua.mrrobot1413.movies.domain.FavoriteRepository
-import ua.mrrobot1413.movies.domain.HomeRepository
+import ua.mrrobot1413.movies.domain.repositories.DetailedRepository
+import ua.mrrobot1413.movies.domain.repositories.FavoriteRepository
+import ua.mrrobot1413.movies.domain.repositories.ReminderRepository
 import javax.inject.Inject
 
 class RemoveFromFavoriteUseCase @Inject constructor(
     private val detailedRepository: DetailedRepository,
-    private val favoriteRepository: FavoriteRepository
+    private val favoriteRepository: FavoriteRepository,
+    private val reminderRepository: ReminderRepository
 ) {
 
     suspend fun invoke(
@@ -22,7 +19,9 @@ class RemoveFromFavoriteUseCase @Inject constructor(
     ) {
         withContext(Dispatchers.IO) {
             detailedRepository.removeFromFavorite(detailedMovieId)
-            favoriteRepository.removeFromFavorite(id)
+            if(!reminderRepository.isToRemindMovie(id)) {
+                favoriteRepository.removeFromFavorite(id)
+            }
         }
     }
 }
